refactor: listen to all localhost addresses
